Added the nestedWhileLoop test case to bound analysis, and the results show that the gradual widening operator can provide more precise results than naive one

+++ tests/bound/nestedwhileloop/naive_bound/nocopy_dealloc/nestedwhileloop.c
@@ -3,7 +3,7 @@ int64_t _f_(int64_t limit){
 	int64_t _1 = 0;
 	int64_t i = 0;
 	int64_t sum = 0;
-	int64_t j = 0;
+	uint16_t j = 0;
 	uint16_t _5 = 0;
 	uint16_t _6 = 0;
 	uint16_t _7 = 0;
